+++ src/components/ng-redux.ts
@@ -18,14 +18,15 @@ import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/filter';
 import 'rxjs/add/operator/distinctUntilChanged';
 import 'rxjs/add/operator/switchMap';
-import { Selector, resolveToFunctionSelector } from './selectors';
+import { Selector, PathSelector, resolveToFunctionSelector } from './selectors';
 import { assert } from '../utils/assert';
+import { SubStore } from './sub-store';
+import { enableFractalReducers } from './fractal-reducer-map';
+import { ObservableStore, Comparator } from './observable-store';
 
-export type Comparator = (x: any, y: any) => boolean;
-
-export class NgRedux<RootState> {
+export class NgRedux<RootState> implements ObservableStore<RootState> {
   /** @hidden */
-  static instance: NgRedux<any> = undefined;
+  static instance: ObservableStore<any> = undefined;
 
   private _store: Store<RootState> = null;
   private _store$: BehaviorSubject<RootState> = null;
@@ -39,7 +40,7 @@ export class NgRedux<RootState> {
   }
 
   /**
-   * configures a Redux store and allows NgRedux to observe and dispatch
+   * Configures a Redux store and allows NgRedux to observe and dispatch
    * to it.
    *
    * This should only be called once for the lifetime of your app, for
@@ -61,7 +62,7 @@ export class NgRedux<RootState> {
     this.setStore(
       compose.apply(null,
         [ applyMiddleware(...middleware), ...enhancers ])(createStore)
-        (rootReducer, initState));
+        (enableFractalReducers(rootReducer), initState));
   }
 
   /**
@@ -79,58 +80,17 @@ export class NgRedux<RootState> {
     this.setStore(store);
   };
 
-  /**
-   * Select a slice of state to expose as an observable.
-   *
-   * @typeparam S
-   * @param selector key or function to select a part of the state
-   * @param [comparator] Optional
-   * comparison function called to test if an item is distinct
-   * from the previous item in the source.
-   *
-   * @returns An Observable that emits items from the
-   * source Observable with distinct values.
-   */
-  select = <S>(
-    selector?: Selector<RootState, S>,
-    comparator?: Comparator): Observable<S> =>
-      this._store$
-        .distinctUntilChanged()
-        .map(resolveToFunctionSelector(selector))
-        .distinctUntilChanged(comparator);
+  // Redux Store methods.
 
-  /**
-   * Get the current state of the application
-   * @returns The application state
-   */
-  getState = (): RootState => this._store.getState()
+  getState = (): RootState =>
+    this._store.getState()
 
-  /**
-   * Subscribe to the Redux store changes
-   *
-   * @param listener A callback to invoke when the state is updated
-   * @returns A function to unsubscribe
-   */
   subscribe = (listener: () => void): Unsubscribe =>
     this._store.subscribe(listener)
 
-  /**
-   * Replaces the reducer currently used by the store to calculate the state.
-   *
-   * You might need this if your app implements code splitting and you want to
-   * load some of the reducers dynamically. You might also need this if you
-   * implement a hot reloading mechanism for Redux.
-   *
-   * @param nextReducer The reducer for the store to use instead.
-   */
   replaceReducer = (nextReducer: Reducer<RootState>) =>
     this._store.replaceReducer(nextReducer)
 
-  /**
-   * Dispatch an action to Redux.
-   *
-   * Ensures that dispatch happens inside the Angular Zone.
-   */
   dispatch: Dispatch<RootState> = action => {
     assert(
       !!this._store,
@@ -140,6 +100,21 @@ export class NgRedux<RootState> {
 
     return this.ngZone.run(() => this._store.dispatch(action));
   };
+
+  // ObservableStore methods
+
+  select = <S>(
+    selector?: Selector<RootState, S>,
+    comparator?: Comparator): Observable<S> =>
+      this._store$
+        .distinctUntilChanged()
+        .map(resolveToFunctionSelector(selector))
+        .distinctUntilChanged(comparator);
+
+  configureSubStore = <SubState>(
+    basePath: PathSelector,
+    localReducer: Reducer<SubState>): ObservableStore<SubState> =>
+      new SubStore<SubState>(this, basePath, localReducer)
 
   private setStore(store: Store<RootState>) {
     this._store = store;