+++ src/index.ts
@@ -1,13 +1,11 @@
-import {
-  NgRedux,
-  Comparator,
-} from './components/ng-redux';
+import { NgRedux } from './components/ng-redux';
 import {
   Selector,
   PathSelector,
   PropertySelector,
   FunctionSelector,
 } from './components/selectors';
+import { ObservableStore, Comparator } from './components/observable-store';
 import { DevToolsExtension } from './components/dev-tools';
 import { select, select$ } from './decorators/select';
 import { dispatch } from './decorators/dispatch';
@@ -29,4 +27,5 @@ export {
   select,
   select$,
   dispatch,
+  ObservableStore,
 };