+++ src/utils/set-in.ts
@@ -1,0 +1,16 @@
+/**
+ * Sets a deeply-nested property value from an object, given a 'path'
+ * of property names or array indices. Path elements are created if
+ * not there already. Does not mutate the given object.
+ *
+ * @hidden
+ */
+export const setIn = (
+  obj: Object,
+  [ firstElem, ...restElems ]: (string | number)[],
+  value): Object => ({
+    ...obj,
+    [firstElem]: restElems.length === 0 ?
+      value :
+      setIn(obj[firstElem] || {}, restElems, value)
+    });