+++ testing/index.ts
@@ -1,6 +1,7 @@
 import { NgReduxTestingModule } from './ng-redux-testing.module';
 import { MockDevToolsExtension } from './dev-tools.mock';
 import { MockNgRedux } from './ng-redux.mock';
+import { MockObservableStore } from './observable-store.mock';
 
 // Warning: don't do this:
 //  export * from './foo'
@@ -10,4 +11,5 @@ export {
   NgReduxTestingModule,
   MockDevToolsExtension,
   MockNgRedux,
+  MockObservableStore,
 };