+++ testing/ng-redux.mock.ts
@@ -2,30 +2,23 @@ import {
   NgRedux,
   Selector,
   Comparator,
+  ObservableStore,
+  PathSelector,
 } from '@angular-redux/store';
+import { Reducer, Action } from 'redux';
 import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
 import { ReplaySubject } from 'rxjs/ReplaySubject';
-
 import 'rxjs/add/observable/from';
 import 'rxjs/add/operator/distinctUntilChanged';
-
-interface SelectorStubRecord {
-  subject: Subject<any>;
-  comparator: Comparator;
-}
-
-interface SelectorStubMap {
-  [selector: string]: SelectorStubRecord;
-}
+import { MockObservableStore } from './observable-store.mock';
 
 /**
  * Convenience mock to make it easier to control selector
  * behaviour in unit tests.
  */
-export class MockNgRedux<RootState> extends NgRedux<RootState> {
-  public static mockInstance: MockNgRedux<any> = null;
-  private static selections: SelectorStubMap = {};
+export class MockNgRedux<RootState> extends MockObservableStore<RootState> {
+  static mockInstance: MockNgRedux<any> = null;
 
   /**
    * Returns a subject that's connected to any observable returned by the
@@ -34,47 +27,39 @@ export class MockNgRedux<RootState> extends NgRedux<RootState> {
    * in the context of a unit test, MockNgRedux will give them the values
    * you pushed onto your stub.
    */
-  public static getSelectorStub<R, S>(
+  static getSelectorStub<R, S>(
     selector?: Selector<R, S>,
     comparator?: Comparator): Subject<S> {
-    return MockNgRedux.initSelectorStub<R, S>(selector, comparator).subject;
+      return MockNgRedux.mockInstance.getSelectorStub<S>(selector, comparator);
+    }
+
+  /**
+   * Returns a mock substore that allows you to set up selectorStubs for
+   * any 'fractal' stores your app creates with NgRedux.configureSubStore.
+   *
+   * If your app creates deeply nested substores from other substores,
+   * pass the chain of pathSelectors in as ordered arguments to mock
+   * the nested substores out.
+   * @param pathSelectors
+   */
+  static getSubStore<S>(...pathSelectors: PathSelector[]): MockObservableStore<S> {
+    return pathSelectors.length ?
+      MockNgRedux.mockInstance.getSubStore(...pathSelectors) :
+      MockNgRedux.mockInstance;
   }
 
   /**
    * Reset all previously configured stubs.
    */
-  public static reset(): void {
-    MockNgRedux.selections = {};
-  }
-
-  private static initSelectorStub<R, S>(
-    selector?: Selector<R, S>,
-    comparator?: Comparator): SelectorStubRecord {
-
-    const key = selector.toString();
-    const record = MockNgRedux.selections[key] || {
-      subject: new ReplaySubject<S>(),
-      comparator,
-    };
-
-    MockNgRedux.selections[key] = record;
-    return record;
+  static reset(): void {
+    MockNgRedux.mockInstance.reset();
   }
 
   /** @hidden */
   constructor() {
-    super(null);
+    super();
 
     NgRedux.instance = this; // This hooks the mock up to @select.
     MockNgRedux.mockInstance = this;
-  }
-
-  public dispatch = () => null;
-
-  public select = <S>(selector?: Selector<RootState, S>, comparator?: Comparator): Observable<any> => {
-    const stub = MockNgRedux.initSelectorStub<RootState, S>(selector, comparator);
-    return stub.comparator ?
-      stub.subject.distinctUntilChanged(stub.comparator) :
-      stub.subject;
   }
 }